Rename file containing transform-related functions
