+++ glowing_waffles/differential_photometry/__init__.py
@@ -1,4 +1,4 @@
 from __future__ import print_function, division, absolute_import
 
-from .filter_transform_ivezic_2007 import *
+from .magnitude_transforms import *
 from .catalog_search import *
