+++ glowing_waffles/differential_photometry/magnitude_transforms.py
